completed implementation of SNV, MNV and Deletion.

+++ src/model/evidence/fragments.rs
@@ -55,14 +55,14 @@ pub fn n_fragment_positions(
 ///
 /// * `left` - left read of the pair
 /// * `right` - right read of the pair
-pub fn estimate_insert_size(left: &bam::Record, right: &bam::Record) -> Result<u32> {
+pub fn estimate_insert_size(left: &bam::Record, right: &bam::Record) -> Result<u64> {
     let left_cigar = left.cigar_cached().unwrap();
     let right_cigar = right.cigar_cached().unwrap();
 
-    let aln = |rec: &bam::Record, cigar| -> Result<(u32, u32)> {
+    let aln = |rec: &bam::Record, cigar| -> Result<(u64, u64)> {
         Ok((
-            (rec.pos() as u32).saturating_sub(evidence::Clips::leading(cigar).both()),
-            cigar.end_pos() as u32 + evidence::Clips::trailing(cigar).both(),
+            (rec.pos() as u64).saturating_sub(evidence::Clips::leading(cigar).both() as u64),
+            cigar.end_pos() as u64 + evidence::Clips::trailing(cigar).both() as u64,
         ))
     };
 
@@ -70,17 +70,17 @@ pub fn estimate_insert_size(left: &bam::Record, right: &bam::Record) -> Result<u
     let (right_start, right_end) = aln(right, &right_cigar)?;
     // as defined by Torsten Seemann
     // (http://thegenomefactory.blogspot.nl/2013/08/paired-end-read-confusion-library.html)
-    let inner_mate_distance = right_start as i32 - left_end as i32;
+    let inner_mate_distance = right_start as i64 - left_end as i64;
 
     let insert_size =
-        inner_mate_distance + (left_end - left_start) as i32 + (right_end - right_start) as i32;
+        inner_mate_distance + (left_end - left_start) as i64 + (right_end - right_start) as i64;
     assert!(
         insert_size > 0,
         "bug: insert size {} is smaller than zero",
         insert_size
     );
 
-    Ok(insert_size as u32)
+    Ok(insert_size as u64)
 }
 
 /// Calculate read evindence for an indel.
